% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta.R
\docType{package}
\name{create.meta}
\alias{create.meta}
\alias{create.meta-package}
\alias{create.metaTree}
\title{Internal function}
\usage{
create.metaTree(depth)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

