% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_max.R
\docType{package}
\name{depth.max}
\alias{depth.max}
\alias{depth.max-package}
\title{Internal function}
\usage{
depth.max(trees)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

