% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Evaluating diversity and plot}
\usage{
diversity(obj, color = NULL, plotit = T)
}
\arguments{
\item{obj}{Object resulted from phyC.}

\item{color}{If plotit=TRUE, color parameter of each cluster in the plot.}

\item{plotit}{Whether plot the diversity or not (logical).}
}
\value{
ind.div Diversity of each tree.

div Diversity of each cluster.
}
\description{
Evaluating diversity and plot
}
\examples{
library(phyC)
data(ccRCC)
vaf <- lapply(vaf,function(x)x[,-(1:3)])
trees <- par.tree(vaf)
edgeList <- lapply(trees,function(x)x$edge)
edgeLenList <- lapply(trees,function(x)x$edge.length)
res <- phyC(edgeList,edgeLenList,cluster = 3,type = "h",method = "ward")
div <- diversity(res)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

