% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{evol}
\alias{evol}
\title{Constructed evolutionary tree set}
\format{A list of 23 edgeList and edgeLenList}
\description{
This data contains the reconstructed evolutionary trees from multiregional sequencing variant allele frequency data set (https://github.com/viq854/lichee/tree/master/LICHeE/data) and (Uchi,et al.2015).
The data set contain 23 patients VAFs consisting the three studies.
}
\section{Variables}{

Variables:
\itemize{
\item edgeList Edge matrix
\item edgeLenList Edge length corresponding to the edge matrix.
}
}
\author{
Yusuke Matsui & Teppei Shimamura
}

