% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\docType{package}
\name{get.index}
\alias{get.index}
\alias{get.index-package}
\title{Internal function}
\usage{
get.index(edge)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

