% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_depth.R
\docType{package}
\name{get.maxdepth}
\alias{get.maxdepth}
\alias{get.maxdepth-package}
\title{Internal function}
\usage{
get.maxdepth(trees)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

