% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichee2edge.R
\name{lichee2edge}
\alias{lichee2edge}
\title{Constructing trees from variant allele frequency using LICHeE.}
\usage{
lichee2edge(licheeDir = NULL, vaf, licheeParamIO = NULL,
  licheeParamFilter = NULL, licheeParamPhy = NULL)
}
\arguments{
\item{licheeDir}{Directory of lichee.jar file.}

\item{vaf}{Matrix of variant allele frequency. For more detail of data format, see http://viq854.github.io/lichee/}

\item{licheeparamIO}{List of input/output and display options.For detail see http://viq854.github.io/lichee/. You can set the parameters of (normal,save,showNetwork,showTree).}

\item{licheeparamFilter}{List of SSNV filtering and calling parameters. For detail see http://viq854.github.io/lichee/. You can set the parameters of (absent,present,maxVAFValid,minProfileSupport).}

\item{licheeparamParamPhy}{List of phylogenetic network construction and tree search paramters. For detail see http://viq854.github.io/lichee/. You can set the parameters of (minClusterSize,minPrivateClusterSize,minRobustNodeSupport,maxClusterDist,completeNetwork,e,nTreeQPCheck).}
}
\value{
edgeList Rooted-constraint network of cancer lineage.

edgeLenList Edge length vector corresponding to the edgeList. The edge legnth means #addition SSNVs from parental clone.
}
\description{
Constructing trees from variant allele frequency using LICHeE.
}
\author{
Yusuke Matsui & Teppei Shimamura
}

