% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regis.R
\docType{package}
\name{meta.regis.tree}
\alias{meta.regis.tree}
\alias{meta.regis.tree-package}
\title{Internal function}
\usage{
meta.regis.tree(meta, target)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

