% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phyC.R
\name{phyC.plot}
\alias{phyC.plot}
\title{Plot trees in clusters}
\usage{
phyC.plot(obj, color = NULL, label = NULL, type = "unrooted")
}
\arguments{
\item{obj}{Object resulted from phyC.}

\item{color}{Vector of color parameter of each cluster in the plot.}

\item{label}{Vector of labels of trees. Default is "Tree i"(i=1,2,...).}

\item{type}{Type of tree to be plotted.  It must be one of "unrooted"(the default), "phylogram", "cladogram", "fan", "radial".}
}
\description{
Plot trees in clusters
}
\examples{
library(phyC)
data(evol)
res <- phyC(evol$edgeList,evol$edgeLenList,cluster=4,type='nh')
phyC.plot(res)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

