% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_tree.R, R/phyC.R
\name{phyC}
\alias{par.tree}
\alias{phyC}
\title{Constructing phylogenetic trees.}
\usage{
par.tree(vaf, thr = 0.05)

phyC(edgeList, edgeLenList, cluster, type = "h", method = "ward")
}
\arguments{
\item{vaf}{Matrix of vaf profie. Each row and column must represent VAFs of genes and sampled region, respectively, and first column must normal cell.}

\item{thr}{Threshold of VAF, which is used for binarizing the VAF profile. If VAF>=thr then it is set to 1; otherwise 0.}

\item{edgeList}{List of edge matrix of input.}

\item{edgeLenList}{List of edge length vector. Index of vector should be corresponding to row index of edge matrix.}

\item{cluster}{The number of clusters.}

\item{type}{Clustering type. Selecht 'nh' (non hierarchical) or 'h'(hierarchical with ward's method).}
}
\value{
tree Constructed phylogenetic trees.

edge Edge matrix of the constructed tree.

edge.length Edge length of the constructed tree.

node.label Labels of each noder representing sub-clones.

profile Mutation status of each sub-clone.

trees Input trees.

regis.tree Registered trees.

cluster Index of the clusters.

dist Distance betwee trees.
}
\description{
Constructing phylogenetic trees.

Clustering cancer evolutionary trees
}
\details{
This function constructsparsimony trees such that nodes and edge length correspond to sub-clones and the number of SSNVs in the sub-clones. acctran in phangorn package is used to construct the parsimony trees.

This function perform the registration and the clustering. In the registration, we resolve the mono- and multi-furcation trees and we complete the number of leaves among the trees. In the resigration, identical tree toplogies are regarded identical even if the label are different. In the clustering, we classify the trees into predifined number of subsets. When choose type='nh', we perform the Ward's clustering.
}
\examples{
library(phyC)
data(ccRCC)
vaf <- lapply(vaf,function(x)x[,-(1:3)])
trees <- par.tree(vaf)
library(phyC)
data(evol)
res <- phyC(evol$edgeList,evol$edgeLenList,cluster=4,type='nh')
res$cluster
}
\author{
Yusuke Matsui & Teppei Shimamura

Yusuke Matsui & Teppei Shimamura
}

