% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phyC.R
\name{plot.phyC}
\alias{plot.phyC}
\title{Plot trees in clusters}
\usage{
\method{plot}{phyC}(obj, color = NULL, label = NULL)
}
\arguments{
\item{obj}{Object resulted from phyC.}

\item{color}{Vector of color parameter of each cluster in the plot.}

\item{label}{Vector of labels of trees. Default is "Tree i"(i=1,2,...).}
}
\description{
Plot trees in clusters
}
\examples{
library(phyC)
##generate edgeList and edgeLenList##
trees <- c(rmtree(5,3),rmtree(5,4))
edgeList <- lapply(trees,function(x)x$edge)
edgeLenList <- lapply(trees,function(x)x$edge.length)
##adopting phyC##
res <- phyC(edgeList,edgeLenList,cluster=2,type='nh')
plot.phyC(res)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

