% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monomulti.R, R/root_find.R
\docType{package}
\name{resolve.monomulti}
\alias{resolve.monomulti}
\alias{resolve.monomulti-package}
\alias{rootFind}
\title{Internal function}
\usage{
resolve.monomulti(edgelist, edgeLenList)

rootFind(edge)
}
\description{
Internal function

Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura

Yusuke Matsui & Teppei Shimamura
}

