% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaf}
\alias{vaf}
\title{Variant allele frequency data (VAF) set}
\format{A list of 14 VAF matrices}
\description{
This data contains the multiregional sequencing variant allele frequency data set from https://github.com/viq854/lichee/tree/master/LICHeE/data.
The data set contain 14 patients VAFs consisting the two studies.
}
\section{Variables}{

Variables:
\itemize{
\item chrom The number of chromosome
\item pos Genomic position
\item DESC Description of the position such as the gene symbol.
\item normal The VAF of the normal cells.
\item Other columns Each column indicates the VAF of each region
}
}
\author{
Yusuke Matsui & Teppei Shimamura
}

