% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichee2edge.R
\docType{package}
\name{lichee2edge}
\alias{lichee2edge}
\alias{lichee2edge-package}
\title{Internal function}
\usage{
lichee2edge(input = NULL, licheeParamIO = NULL, licheeParamFilter = NULL,
  licheeParamPhy = NULL, outdir = NULL, idx = 1)
}
\description{
Internal function
}
\author{
Yusuke Matsui & Teppei Shimamura
}

