% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyC.R
\docType{package}
\name{phyC}
\alias{phyC}
\alias{phyC-package}
\title{Clustering cancer evolutionary trees}
\usage{
phyC(edgeList, edgeLenlist, cluster, type = "nh", method = NULL)
}
\arguments{
\item{edgeList}{List of edge matrix that include the root.}

\item{cluster}{The number of clusters.}

\item{type}{Clustering type. Selecht 'nh' (non hierarchical) or 'h'(hierarchical with ward's method).}

\item{edgeLenList}{List of edge length vector. Index of vector should be corresponding to row index of edge matrix.}

\item{edgeList}{List of edge matrix of input.}

\item{edgeLenList}{List of edge length vector of input.}
}
\value{
trees Input trees.

regis.tree Registered trees.

cluster Index of the clusters.

dist Distance betwee trees.
}
\description{
Clustering cancer evolutionary trees
}
\details{
This function perform the registration and the clustering. In the registration, we resolve the mono- and multi-furcation trees and we complete the number of leaves among the trees. In the resigration, identical tree toplogies are regarded identical even if the label are different. In the clustering, we classify the trees into predifined number of subsets. When choose type='nh', we perform the Ward's clustering.
}
\examples{
library(phyC)
##generate edgeList and edgeLenList##
trees <- c(rmtree(5,3),rmtree(5,4))
edgeList <- lapply(trees,function(x)x$edge)
edgeLenList <- lapply(trees,function(x)x$edge.length)
##adopting phyC##
res <- phyC(edgeList,edgeLenList,cluster=2,type='nh')
}
\author{
Yusuke Matsui & Teppei Shimamura
}

