% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyCMD.R
\name{phyCMD}
\alias{phyCMD}
\title{Constructing configuration of trees in clusters and plot}
\usage{
phyCMD(obj, color = NULL, label = NULL, img.width = 200,
  img.height = 200, size = 0.15, type = "unrooted")
}
\arguments{
\item{obj}{Object resulted from phyC.}

\item{color}{Vector of color parameter of each cluster in the plot.}

\item{label}{Vector of labels of trees. Default is "Tree i"(i=1,2,...).}

\item{size}{Size parameter of trees to be plotted in the Euclidean space. Usually, 0 < size < 1. Default is 0.15.}

\item{type}{Type of tree to be plotted.  It must be one of "unrooted"(the default), "phylogram", "cladogram", "fan", "radial".}

\item{img.width(/img.height)}{Image width (/ height) of the trees to be overlayed on the Euclidean space. The unit is "px". Default is 200.}
}
\value{
dist Distance used for configuration

coord Coordinate of trees in the Euclidean space.
}
\description{
Constructing configuration of trees in clusters and plot
}
\details{
This function performs classical multidimensional scaling with tree distance. The resulting plot includes the trees overlayed on the Euclidean coordinates.
}
\examples{
library(phyC)
data(evol)
res <- phyC(evol$edgeList,evol$edgeLenList,cluster=4,type='nh')
phyCMD(res)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

